# api/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API Gateway of Personalized Medicine and Health Monitoring Platform",
        "title": "Personalized Medicine and Health Monitoring Platform",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/customer/lifestyle": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new lifestyle data record for a user",
                "tags": [
                    "lifestyle"
                ],
                "summary": "Adds new lifestyle data",
                "parameters": [
                    {
                        "description": "Lifestyle data",
                        "name": "lifestyle_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/health_analytics.AddLifestyleDataRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/health_analytics.AddLifestyleDataResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while processing request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/customer/lifestyle/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves lifestyle data by ID",
                "tags": [
                    "lifestyle"
                ],
                "summary": "Gets lifestyle data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lifestyle data ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/health_analytics.GetLifestyleDataResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid data ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while processing request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates existing lifestyle data",
                "tags": [
                    "lifestyle"
                ],
                "summary": "Updates lifestyle data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lifestyle data ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated lifestyle data",
                        "name": "lifestyle_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/health_analytics.UpdateLifestyleDataRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/health_analytics.UpdateLifestyleDataResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while processing request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes lifestyle data by ID",
                "tags": [
                    "lifestyle"
                ],
                "summary": "Deletes lifestyle data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lifestyle data ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/health_analytics.DeleteLifestyleDataResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid data ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while processing request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "health_analytics.AddLifestyleDataRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/health_analytics.LifestyleData"
                }
            }
        },
        "health_analytics.AddLifestyleDataResponse": {
            "type": "object",
            "properties": {
                "data_id": {
                    "type": "string"
                }
            }
        },
...
        "health_analytics.DeleteLifestyleDataResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "health_analytics.GetLifestyleDataResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/health_analytics.LifestyleData"
                }
            }
        },
...
        "health_analytics.LifestyleData": {
            "type": "object",
            "properties": {
                "data_type": {
                    "type": "string"
                },
                "data_value": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "string"
                },
                "recorded_date": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "health_analytics.UpdateLifestyleDataRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/health_analytics.LifestyleData"
                }
            }
        },
        "health_analytics.UpdateLifestyleDataResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}